# mnemonic/internal/http/ui/openapi.yaml
openapi: 3.0.3
info:
  title: Skillgrid Mnemonic HTTP API
  description: |-
    Persistent memory, code index, and web research cache for AI agents.

    Write routes (POST) optionally require an `Authorization: Bearer <token>`
    header when the server is started with `SKILLGRID_HTTP_TOKEN` set.
    Most routes take a `project` query parameter naming the project store
    (see `GET /projects` for available IDs).
  version: "0.1.0"
servers:
  - url: /
...
paths:
...
  /sessions/{id}/tool-calls:
    post:
      tags: [session]
      operationId: sessionToolCall
      summary: Record one harness tool call
      description: |-
        Posted by the Cursor, OpenCode, and Kilo hooks after every tool call.
        Creates the session (tagged with agent) when it does not exist yet.
        Bumps the session's read, write, command, and error counters and
        appears on GET /activity/stream as a `tool` frame.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: project, in: query, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                agent: { type: string, example: cursor }
                directory: { type: string }
                type: { type: string, description: "file_read, file_write, shell, or other" }
                tool_name: { type: string }
                path: { type: string }
                command: { type: string }
                result_status: { type: string, example: success }
                content_hash: { type: string }
                content_preview: { type: string }
      responses:
        "200": { description: Recorded }
        "400": { $ref: "#/components/responses/Bad" }
...
  /context:
    get:
      tags: [memory]
      operationId: context
      summary: Recent session context
      parameters:
        - name: project
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Recent sessions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sessions"
        "400":
          $ref: "#/components/responses/Bad"
        "500":
          $ref: "#/components/responses/Internal"
...
components:
  responses:
    Bad:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPError"
    Internal:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPError"
...
  schemas:
...
    HTTPError:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Sessions:
      type: object
      required: [sessions]
      properties:
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/Session"
...
    Session:
      type: object
      properties:
        id:
          type: string
        project:
          type: string
        directory:
          type: string
        title:
          type: string
          nullable: true
          description: Human-readable session name (optional); shown in the dashboard session list
        started_at:
          type: string
        ended_at:
          type: string
        summary:
          type: string
        status:
          type: string
...